requirements.txt
--index-url https://download.pytorch.org/whl/cu111
torch==1.8.0+cu111
torchvision==0.9.0
numpy==1.24.3
pillow==10.0.1
matplotlib==3.7.4
tqdm==4.66.1
torchsummary
pyyaml==6.0.1